grocery store software.py
This is a small grocery store software. it is in python. Software in which, when the digits or a specific number is entered, then the some action occurs, and also it generates the auto bill for it. So this is the source code. You can also make changes in this source code. So follow my GITHUB profile. Thank you.